Move ValueType to Concept
