+++ api/concept/type/Type.java
@@ -26,15 +26,10 @@ import com.eclipsesource.json.JsonObject;
 import com.vaticle.typedb.client.api.TypeDBTransaction;
 import com.vaticle.typedb.client.api.concept.Concept;
 import com.vaticle.typedb.client.common.Label;
-import com.vaticle.typedb.client.common.exception.TypeDBClientException;
-import com.vaticle.typedb.protocol.ConceptProto;
 
 import javax.annotation.CheckReturnValue;
 import javax.annotation.Nullable;
-import java.time.LocalDateTime;
 import java.util.stream.Stream;
-
-import static com.vaticle.typedb.client.common.exception.ErrorMessage.Concept.BAD_VALUE_TYPE;
 
 public interface Type extends Concept {
 
@@ -77,74 +72,5 @@ public interface Type extends Concept {
 
         @CheckReturnValue
         Stream<? extends Type> getSubtypesExplicit();
-    }
-
-    enum ValueType {
-
-        OBJECT(Object.class),
-        BOOLEAN(AttributeType.Boolean.class),
-        LONG(AttributeType.Long.class),
-        DOUBLE(AttributeType.Double.class),
-        STRING(AttributeType.String.class),
-        DATETIME(LocalDateTime.class);
-
-        private final Class<?> valueClass;
-
-        ValueType(Class<?> valueClass) {
-            this.valueClass = valueClass;
-        }
-
-        @CheckReturnValue
-        public static ValueType of(ConceptProto.ValueType valueType) {
-            switch (valueType) {
-                case BOOLEAN:
-                    return BOOLEAN;
-                case LONG:
-                    return LONG;
-                case DOUBLE:
-                    return DOUBLE;
-                case STRING:
-                    return STRING;
-                case DATETIME:
-                    return DATETIME;
-                default:
-                    throw new TypeDBClientException(BAD_VALUE_TYPE, valueType);
-            }
-        }
-
-        @CheckReturnValue
-        public static ValueType of(Class<?> valueClass) {
-            for (ValueType t : ValueType.values()) {
-                if (t.valueClass == valueClass) {
-                    return t;
-                }
-            }
-            throw new TypeDBClientException(BAD_VALUE_TYPE);
-        }
-
-        @CheckReturnValue
-        public Class<?> valueClass() {
-            return valueClass;
-        }
-
-        @CheckReturnValue
-        public ConceptProto.ValueType proto() {
-            switch (this) {
-                case OBJECT:
-                    return ConceptProto.ValueType.OBJECT;
-                case BOOLEAN:
-                    return ConceptProto.ValueType.BOOLEAN;
-                case LONG:
-                    return ConceptProto.ValueType.LONG;
-                case DOUBLE:
-                    return ConceptProto.ValueType.DOUBLE;
-                case STRING:
-                    return ConceptProto.ValueType.STRING;
-                case DATETIME:
-                    return ConceptProto.ValueType.DATETIME;
-                default:
-                    return ConceptProto.ValueType.UNRECOGNIZED;
-            }
-        }
     }
 }
